Major update: change several functions and classes for a new (and simpler) way to get the model.
I create the functions 'fit', 'predict' and 'score' and I adapted the class and methods to a new way to train e obtain a model. The class has now a default behavior. Now the model can be trained using only 'clf = sys_identfy().fit(X, y)'. The prediction and scoring can be done in a similar form.

Also, I changed the name of some variables and some minor changes in functions. There are other minor changes in the code, but not worth to mention in detail here.